Only update vrs sp if not popped from stack
When executing the 'Pop up to 12 integer registers under masks'
instruction it is possible for the mask to specify that register 13, the
stack pointer, is to be popped. The ARM documentation (Exception
handling ABI for the ARM architecture, section 9.3, remark b) specifies
that the update to the vrs r13 must be delayed until execution of the
whole instruction has completed. Since this implementation increments a
copy of vrs r13 ('vsp' variable) during the execution of an instruction
it is simply a case of not updating vrs r13 with this copy if r13 was
included in the pop mask.