Fix implementation of VFP instruction execution
I'm not sure what the code was trying to do, but it did not match the
ARM documentation at all. The 2-byte instructions did not get the second
byte of the instruction, the vrs stack pointer was never incremented,
the number of VFP registers being popped was not being used, the +4
increment was not being applied for the 'as if by FSTMFDX' variants and
the vrs r14 (lr) register was being fiddled with. This fixes all of
these things.